requirements.txt
django == 3.1.2
pillow
django-crispy-forms
django-rest-framework
django-admin-interface
django-filters
coverage
django-silk